chore: sync live metric tiles pinned row with data row updates
